Moved restricted code to separate file.
